Don't always convert from json string to edn
Motivation:

With the swagger lib, the reply now does not return a json string, but
actual edn.

Modifications:

Create a fn that is smart enough on how to convert to edn. If the param
is already a clojure object return it. If it's a string, build as json.
Else, it may be a stream that will handle a string, so slurp it and
build as json.